SHOPRunFilter class implementation
